+++ components/features/apm/services/[serviceId]/logs/LogItem.tsx
@@ -1,6 +1,6 @@
 /// 에러 로그 1개 => UI 스타일만 담당
 
-import type { LogEntry } from './types';
+import type { LogEntry } from '../../../../../../types/LogsTypes';
 import LevelBadge from './LevelBadge';
 
 export default function LogItem({ item }: { item: LogEntry }) {
