+++ components/ui/StatGrid.tsx
@@ -1,6 +1,6 @@
 // 에러 카드를 모아서 보여줌 => 통계 데이터 묶음 + 레이아웃 담당
 
-import type { StatItem } from '../features/apm/services/[serviceId]/logs/types';
+import type { StatItem } from '../../types/LogsTypes';
 import StatCard from './StatCard';
 
 export default function StatGrid({ items }: { items: StatItem[] }) {
