+++ types/LogsTypes.ts
